youtube_bot/requirements.txt
aiogram==3.15.0
yt-dlp==2024.12.23
python-dotenv==1.0.0
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0
asyncpg==0.29.0
redis==5.0.1
